Insert some casts to eliminate some warning messages from GCC.
This was contributed by DanGrayson as a pull request
to the a-star repository:
https://github.com/pololu/a-star/pull/1

+++ FastGPIO/FastGPIO.h
@@ -71,17 +71,17 @@ namespace FastGPIO
 
         volatile uint8_t * pin() const
         {
-            return (volatile uint8_t *)pinAddr;
+            return (volatile uint8_t *)(uint16_t)pinAddr;
         }
 
         volatile uint8_t * port() const
         {
-            return (volatile uint8_t *)portAddr;
+            return (volatile uint8_t *)(uint16_t)portAddr;
         }
 
         volatile uint8_t * ddr() const
         {
-            return (volatile uint8_t *)ddrAddr;
+            return (volatile uint8_t *)(uint16_t)ddrAddr;
         }
     } IOStruct;
     /** @endcond */
